+++ client/src/main/java/com/chencraft/ntu/model/response/IntResponse.java
@@ -6,5 +6,5 @@ import lombok.experimental.SuperBuilder;
 @Getter
 @SuperBuilder
 public class IntResponse extends GenericResponse {
-    private long value;
+    private int value;
 }
